chore(core): remove duplicate mergeVersions
