+++ .changeset/hot-ads-notice.md
@@ -1,0 +1,6 @@
+---
+'@verdaccio/core': patch
+'@verdaccio/store': patch
+---
+
+chore(core): remove duplicate mergeVersions
