+++ packages/core/core/src/pkg-utils.ts
@@ -50,37 +50,6 @@ export function getLatest(pkg: Manifest): string {
 }
 
 /**
- * Function gets a local info and an info from uplinks and tries to merge it
- exported for unit tests only.
-  * @param {*} local
-  * @param {*} upstream
-  * @param {*} config sds
-  * @deprecated use @verdaccio/storage mergeVersions method
-  */
-// @deprecated
-export function mergeVersions(local: Manifest, upstream: Manifest) {
-  // copy new versions to a cache
-  // NOTE: if a certain version was updated, we can't refresh it reliably
-  for (const i in upstream.versions) {
-    if (typeof local.versions[i] === 'undefined') {
-      local.versions[i] = upstream.versions[i];
-    }
-  }
-
-  for (const i in upstream[DIST_TAGS]) {
-    if (local[DIST_TAGS][i] !== upstream[DIST_TAGS][i]) {
-      if (!local[DIST_TAGS][i] || semver.lte(local[DIST_TAGS][i], upstream[DIST_TAGS][i])) {
-        local[DIST_TAGS][i] = upstream[DIST_TAGS][i];
-      }
-      if (i === 'latest' && local[DIST_TAGS][i] === upstream[DIST_TAGS][i]) {
-        // if remote has more fresh package, we should borrow its readme
-        local.readme = upstream.readme;
-      }
-    }
-  }
-}
-
-/**
  * Get package informations from package.json
  * @param packagePath - The path to the package.json file.
  * @returns The package.json content
