+++ packages/core/core/test/pkg-utils.spec.ts
@@ -26,6 +26,15 @@ describe('pkg-utils', () => {
     ).toBe('1.0.0');
   });
 
+  test('semverSort', () => {
+    expect(pkgUtils.semverSort(['1.2.3', '1.2', '1.2.3a', '1.2.3c', '1.2.3-b'])).toStrictEqual([
+      '1.2.3a',
+      '1.2.3-b',
+      '1.2.3c',
+      '1.2.3',
+    ]);
+  });
+
   test('semverSanitize', () => {
     expect(pkgUtils.semverSanitize('1.0.0')).toBe('1.0.0');
     expect(pkgUtils.semverSanitize('1.0.0-beta.1')).toBe('1.0.0-beta.1');
